factshield/requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
torch
pandas
scikit-learn
tiktoken
requests
feedparser
beautifulsoup4
langdetect
deep-translator
datasets
accelerate
googlesearch-python
newspaper3k
lxml_html_clean
python-dotenv
emoji